change paf_file_index type from u32 to u16 to reduce file index size

+++ src/impg.rs
@@ -86,7 +86,7 @@ pub struct QueryMetadata {
     target_end: i32,
     query_start: i32,
     query_end: i32,
-    paf_file_index: u32,          // Track which PAF file this record belongs to
+    paf_file_index: u16, // Track which PAF file this record belongs to (max 65535 files)
     strand_and_cigar_offset: u64, // Track strand and cigar offset
     cigar_bytes: usize,
 }
@@ -332,7 +332,7 @@ impl Impg {
                             target_end: record.target_end as i32,
                             query_start: record.query_start as i32,
                             query_end: record.query_end as i32,
-                            paf_file_index: file_index as u32,
+                            paf_file_index: file_index as u16,
                             strand_and_cigar_offset: record.strand_and_cigar_offset, // Already includes strand bit
                             cigar_bytes: record.cigar_bytes,
                         };
